refactor: modularize the code and simplified index.tsx
- Extracted UI components (SingleSelectButtonGroup, MultiSelectButtonGroup, Slider)
- Separated generator logic into PasswordGenerator and PassphraseGenerator
- Created utility functions for clipboard and color scheme detection
- Created dedicated components for PassphraseDisplay, SecurityInfo, and Footer
- Separated password and passphrase settings into distinct components
- Centralized type definitions in generator.ts
