+++ types/generator.ts
@@ -1,0 +1,5 @@
+export type GeneratorMode = "basic" | "advanced";
+export type SeparatorOption = "-" | "_" | "." | "+";
+export type GeneratorType = "password" | "passphrase";
+export type SymbolOption = "#" | "*" | "!" | "%";
+export type ButtonSize = "small" | "medium";
